DataStore architecture beginnings
Starting to pay a little more attention to how fetching data will work,
as well as how to move data around the app. I’m intentionally doing
things in a “flux-ish” style without using a library. I want to really
feel the pain to motivate bringing on a new dependency.